Remove implicit declaration of strnlen.
When running make a variety of errors and warnings occur because of the
declaration of strnlen in gbuf.c. Commenting it out seems to fix the
build and has yet to affect the running of rememberance agent.

```
gbuf.c:43:8: error: conflicting types for ‘strnlen’
 size_t strnlen (char *s, size_t n)
        ^
In file included from ../savant.h:50:0,
                 from gbuf.c:39:
/usr/include/string.h:406:15: note: previous declaration of ‘strnlen’ was here
 extern size_t strnlen (const char *__string, size_t __maxlen)
               ^
```